refactor: extract getRouteEntry and update header
